Issue #12245: Remove extended project browser feature
The extended project browser feature consists of two combo boxes (one
for the tier 1 projects and one for the tier 2 projects). Extensive
JavaScript is used to link the two combo boxes together so that changing
the tier 1 project refreshes the tier 2 project list.

This approach suffers from a number of issues:
1) It doesn't escape potentially harmful values before inserting them
within the JavaScript.
2) It uses inline JavaScript which we're avoiding due to #11826.
3) It doesn't scale beyond 2 levels of projects (project and
sub-projects).
4) When enabled there is no fallback mechanism for JavaScript-disabled
browsers.

Therefore this feature should be deprecated, removed and replaced in the
future with a superior and more modern alternative.