Fix #12249: g_error_send_page_header condition check always true
Within the error_handler() function of error_api.php, this check was
being performed:

if( $g_error_send_page_header || $g_error_send_page_header == null )

Due to PHP's handling of boolean logic this check would always be
validated as true. To solve this problem we simplify the evaluation by
removing the possibility that the variable can be null (we set it to
true by default now).