Fix #11983: Excel export crashes due to missing due date export function
The excel_format_due_date() function was missing from excel_api.php.
When the due date column was turned on and the Excel export feature
used, the export would crash due to not being able to find the missing
due date export function.

Thanks to Dominik Blunk for this patch.