Fix #11758: Improve handling of reassign on feedback feature
The bugnote_add() function in bugnote_api.php is currently the only way
(using the API) to add a bugnote into the database. This function was
also attempting to handle the reassign on feedback feature. This posed
a problem because there are times that the bugnote_add() function needs
to be called without considering the reassign on feedback feature. For
example, importing issues from an XML file.

The bugnote_add() function has been changed to no longer handle the
reassign on feedback feature. The resposibility for handling this
feature now belongs within the scripts calling the bugnote_add()
function. The caller scripts are in a much better position to determine
how the reassign on feedback issue will be handled.