Diagnostic Logging
This is hopefully a simplification for the end user of the diagnostic logging setup.

There is now a log_destination of 'page' - which display diagnostic output at the end of the page, replacing the queries_list functionality - therefore 'show_queries_list' has been removed.

Equivalent output can be achieved by setting:

$g_log_level = LOG_DATABASE;
$g_log_destination = 'page';

In addition, a log_destination of 'none' has been added such that a plugin can be used to handle events without events going to php's error_log.

For production use, an admin check has been added suggesting that page|firebug are not the destination, and warning if diagnostic logging is on.

Note: Currently unique query count is only shown with queries_count if log_level=DB (log_destination=none).