Fix #12327 - Enhance plugin filters to allow developers to specify the number of columns to use in the bug filter.
Feature allowing plugin filters to specify the number of table columns to
span in the bug filter.  If a field requires more columns than are available
in the row it is placed into the next row with available columns.  If the
plugin specifies a colspan greater than is configured the field is given a
full row.  If the plugin specifies a colspan of zero the field is assigned
one column.  If the plugin specifies a negative number it is converted to a
positive number.