Fix #7328: ini_get_number uses wrong postfix multipliers
The ini_get_number function currently uses the wrong postfix multipliers
when determining the integer value of configuration options from
php.ini.

The PHP manual clearly states that k = kibibyte instead of k = kilobyte
as per
http://www.php.net/manual/en/faq.using.php#faq.using.shorthandbytes

Thus we need to change the postfix multipliers to reflect what the PHP
manual defines for php.ini.

Thanks to Morgan Parry for providing a patch to this issue just over 4
years ago. And thanks to Roland Becker for bumping up old issues like
this one that have been forgotten in the past.