Fix #9663: Roadmap underlines incorrect lengths
The text underlines on the roadmap page are displayed with an incorrect
length when the version name contains characters that are encodable via
htmlspecialchars(). We need to use the length of the raw string instead
of the length of the encoded string.

Thanks to Roland Becker for the initial patch for this issue.