Fix #10995: Correct font paths and error messages for ezGraph
We need to return errors as an image so that they appear within an <img>
tag within HTML. Furthermore, the function needs to return the full path
to the font file in use. This is in contrast to the JpGraph mode where
we only need to return the file name of the font file. This discrepancy
would be good to avoid in the future if possible.