Fix #11351: Do not delete email or realname when editing user with LDAP
When connecting to Mantis with LDAP and either use_ldap_email or
use_ldap_realname = ON, that field is set to blank when the user edits
their profile (account_page.php). The same happens when either or both
of the above options are ON and the administrator updates a user from
manage_user_edit_page.php.

The original patch from Damien was updated to fix a few minor bugs and
more importantly, to resolve a number of potential XSS vulnerabilities.

Co-contributed-by: David Hicks <hickseydr@optusnet.com.au>
Signed-off-by: David Hicks <hickseydr@optusnet.com.au>