Fix #12167: Improve LDAP logging and comments in config_defaults.php
Document the fact that LDAP port parameter is not used by ldap_connect when the provided hostname is a URI, and modify the logging in ldap_api.php to correctly reflect what is actually happening to avoid creating confusion.

Implemented also additional improvements to LDAP logging, allowing to fully trace what is happening throughout the LDAP authentication process.

Signed-off-by: David Hicks <d@hx.id.au>