Fix #12363: Billing report end date is inaccurate
The strototime() function was being called incorrectly, leading to the
last day of a billing report range not being included in the results.

There was also an off-by-one-second mistake with calculating the
timestamp denoting the end of any given day in the billing report range.

Thanks to vr5 for the initial patch.