Issue #12545: Enforce XHTML strict output for all pages served
We should be using XHTML strict for the pages served up by MantisBT so
that browsers can quickly ensure that the MantisBT pages are valid/don't
contain errors.

This will require quite a fair bit of trial and error testing from the
developer end to ensure that the XHTML markup is valid on all pages.

Note: this WILL cause breakages - it's meant to. This is the only way to
really discover errors in the output we're sending to clients. Plugins
are affected too and need to be sending XHTML strict output as well.