Fix #12545: Refactor print_timezone_option_list()
This function was not producing well formed XHTML due to missing closing
tags on the <optgroup> elements. It was also neglecting to send line
breaks and tabs to the output engine for each option printed, causing
very long lines to appear in the XHTML output. Shorter lines are easier
to read through and debug.