Fix #12545: Refactor manage_proj_edit_page to be XHTML well formed
There were a number of errors in the output from
manage_proj_edit_page.php which led to XHTML parsing errors. A </form>
closing tag was in the wrong place and and <input> tag was unclosed.

Some minor reformatting/spacing of sections of this file has also been
conducted to improve readability.