Fix #11909: history_localize_item sending wrong argument type to bug_revision_exists
Within history_api, the history_localize_item function is calling the
bug_revision_exists function of bug_revision_api with the wrong argument
type. It should be sending an integer, not a string.

As bug_revision_api doesn't use db_prepare_int when building queries
this error will result in SQL query execution errors when an integer
field in the database is compared to the supplied string (type
mismatch).