Issue #12545: Replace deprecated <b> tags with <strong>
The old HTML <b> tag is deprecated and can be replaced with the <strong>
tag instead. Note that this replacement is not always true - <strong> is
meant to convey contextual meaning rather than just bold text style. For
now we'll just perform the replacement without consideration of whether
the usage of <strong> makes contextual sense in each usage scenario.