Fix #12641: Internet Explorer does not support xhtml+xml MIME type
Internet Explorer prior to version 9 does not support the
application/xhtml+xml MIME type and will therefore not render pages
served up in this way.

We should check the "Accept" header sent by the browser to ensure that
the browser can handle application/xhtml+xml before we use this MIME
type. Otherwise the default should be to use the text/html MIME type for
older browsers.

Signed-off-by: David Hicks <hickseydr@optusnet.com.au>