Issue #12619: Custom fields are reported as required after upgrade from 1.1.0 to 1.2.3. Issue #11684: Incorrect error "A necessary field "MyField" was empty. Please recheck your inputs." when submitting new issue.
There is a lot of details about this issue in issue #11684.  In 1.2.1 a breaking change was introduced which blocks users from submitting / updating issues for projects with certain custom field configurations.  The aim of this fix is to unblock these users.  We can do follow up fixes to improve on this for 1.2.x or to come up with the alternative strategy for 1.3.x.  The goal is to come up with an approach that doesn't break existing users.  The worst thing is that users upgrade and then find out that they can continue to use the product.  If there is a path for upgrade to avoid these issues, then we should automate it as part of the upgrade script.

I haven't marked the issue as completely fixed until dhx checks it out, since he was working on the issue.  I've used my fix for #12619 + one more fix for the update code path.  When I compared my fix to the attached patches, I note that the patches have more changes to handle things like special cases for checkboxes and the case where user doesn't have read-write.  We may incorporate these after we do the necessary validations.  Thanks for Sergiodf for contributing the patches.