Fix #11687: Bugs with attachments that are moved will lose attachments
When a bug is logged and assigned to a project which has a project path
assigned, and later a mantis "administrator" moves the bug from the
initial project to a new project, the attachments are lost.

Jacob's initial patch from mid 2010 has been updated to work with the
latest version of MantisBT 1.3.x. Furthermore the patch has been cleaned
up and simplified. Needs testing however!

Signed-off-by: David Hicks <hickseydr@optusnet.com.au>