Deprecate $g_use_iis as it no longer applies to newer versions of IIS
$g_use_iis was used years ago in MantisBT to work around some bugs in
earlier versions of IIS (such as v5). This is no longer an issue for the
1.3.x branch because it is expected that users of MantisBT will not be
using ancient versions of IIS.

The Refresh header is not needed: we should be using Location instead.
This will automatically add a 302 redirect status code to the HTTP
header and therefore we don't need to supply the code ourselves. PHP
output gzip compression should work fine with IIS if users setup their
servers correctly.