Remove compressed MantisBT JavaScript
The two versions (development and 'production') are out of sync with
each other. As stated before, it is better for us to bundle compressed
'production' versions of the MantisBT JavaScript files when we're
preparing the release packages. This will avoid the issue of JavaScript
files becoming out of sync in the repository (not a good situation).