Bump minimum PHP requirement from 5.3.0 to 5.3.2
The earliest versions of the 5.3 branch of PHP are riddled with
showstopper bugs and therefore cannot be used reliably with MantisBT.

Version 5.3.2 is the lowest version included by the major Linux
distributions (RHEL6 includes 5.3.2) and therefore should be able to
used as the minimum requirement without too much concern.

This requirement change should reduce the number of bugs being reported
to mantisbt.org/bugs that are PHP bugs rather than MantisBT bugs.

The minimum installation environment documentation has also been
updated and clarified.