Fix #12884: Checks for obsolete config fails to report error
Function config_obsolete in core/config_api.php failed to report deprecated
configs defined in core/obsolete.php and displaying replacement parameter,
due to inverted parameters in call to print_test_warn_row.

Also fixes an error when calling the function with a single parameter (as
is sometimes the case in obsolete.php) by making the 2nd parameter optional
and defaulting to ''.

Signed-off-by: David Hicks <d@hx.id.au>