Fix #12675: Do not attempt connection to non-existent database
The original fix for #12675 was incorrect - the patch was applied to the
wrong line of install.php. We should not be attempting to connect to a
specific database that doesn't exist and needs creating.

The proper solution to #12675 requires more thought and work because
only certain database backends support connections to database servers
without specifying a database to connect to.