Fix #12794 : Java Exception when the Issue has an attached file (using mantis plugin in Hudson/Jenkins)
Do not use DIRECTORY_SEPARATOR to build paths which will be converted
to URLs. The main usage of DIRECTORY_SEPARATOR is to convert paths
which are received from the operating system, which is not the case
when determining the application's install location.

The slash is a valid separator on all PHP platforms [1], therefore
we use that.

  [1]: http://php.net/manual/en/function.basename.php