SOAP API: Issue updates only set the provided fields
When updating an issue the issue is first loaded from the database and
then only the provided fields are used to update it. The sponsorship
total is no longer updated since it is a computed field and should not
be manipulated directly.

Fixes #10323: mc_issue_update() will set optional fields to default
values rather than leave them with current values