Initialize the user summary array before calculating the billing total
When calculating the user's total time, the use of += operator in the for loop
causes the display of an E_NOTICE system error, due to access of an undefined
index during the first assignment for a given user.

This has no impact when using default values for $g_display_errors since
E_NOTICE is set to none.