Fix #12998: Password reset button should be hidden for HTTP_AUTH method
If MantisBT is configured with the HTTP_AUTH method for user
authentication, passwords cannot be reset by MantisBT. The passwords are
managed externally to MantisBT in this instance.

Therefore do not offer managers the ability to reset passwords when
HTTP_AUTH is set as the authentication mechanism.

Thanks to Tino Böhme for reporting this issue and providing a patch.