Remove trailing colon on field labels
In manage_user_edit_page.php, the use of a trailing colon on field labels is
not consistent. Since other edit pages in MantisBT do not display colons after
field labels, they have been removed from this page too.

David Hicks mentioned that "these characters are actually styling marks and
therefore are better applied using the 'after' property in CSS."

Note: this is a manual re-coding of what was done in commit a48fd743e, needed
as in 1.3 the colons were moved from the code to language strings.