Use placeholder attribute for stopwatch input fields
As part of issue #11826 to remove all inline JavaScript, the time
tracking stopwatch feature was rewritten using jQuery. A default value
of "hh:mm" was being placed in the time tracking input field which is
invalid if submitted.

We should be using the newer 'placeholder' HTML attribute instead of
mangling the actual value stored in the text input field.