Use placeholder attribute for stopwatch on bug update page
This is a follow up to commit 10af3a5dacc91c3e5332. Use the new HTML
placeholder attribute to provide a formatting hint (hh:mm) instead of
making this hint part of the field label. This also makes it easy to
edit the field - click and type, instead of having to delete a default
value.