Fix #13163: Remove limitation on password length with MD5 authentication
A new function auth_get_password_max_size was added in authentication_api.php,
to return the maximum length of the password, taking the login method into
consideration: limited to the database field size (PASSLEN) for PLAIN and
BASIC_AUTH, or to new constant MAX_PASSWORD_SIZE for other, hash-based methods.

The return value is used to define the maxlength attribute of all the password
fields.