Add new database function to check identifier size
The code fixing #13227 was modified, to remove the str_replace that deletes the
the hardcoded "_table" suffix from the table name. We rely instead on the user
modifying the db prefix and suffix in their config_inc.php to keep them short
enough (or even blank) so that we don't go over the 30 characters limit in
Oracle. This is cleaner than arbitrarily changing the table in the code.

Porting to 1.3 - Conflicts:
	core/database_api.php
	lang/strings_english.txt