Correct naming of attachments column
Commit 97b67e4bbb2f2df5f6fb24eea5b4bb37e1c28acd introduced some changes to the
BugData class: the 'attachment' column was renamed to 'attachment_count'. This
change was not cascaded everywhere, which resulted in the column not being
displayed by default (as the now-invalid column named 'attachment' was
automatically removed)

This then hid the fact that $g_show_attachment_indicator was actually not
working as advertised, i.e. the attachments column is always displayed; when
ON the field contains the paperclip bitmap, and when OFF the actual number of
attachments for the corresponding issue.

Affects issue #13276

This is porting d86ba7220b1b5f9b6eab2eaef44f1345e6699df0 to master