Remove use of filter_var(...) for master-1.2.x
Robert Munteanu correctly mentioned on the mailing list that the
master-1.2.x branch needs to work with PHP 5.1. The filter_var(...)
function is not available in PHP <5.2 and therefore we can't use it in
the master-1.2.x branch.

We can achieve similar results by removing all null characters from the
string (\0). This is OK for our use case which is the use of
$_SERVER['SCRIPT_NAME'] - a value that is set by the server and cannot
be manipulated by remote users (like PHP_SELF).