Fix #13303: use strtolower in ini_get_bool()
Generally, ini_get() returns '0' or '1' when querying a boolean option,
regardless of the actual string used to set it (i.e. On, False, etc).
However, in some cases (e.g. when the option is set in httpd.conf
via php_admin_value instead of php_admin_flag), ini_get() returns the
string itself instead.

This caused ini_get_bool() to wrongly return False when the option is
set to an equivalent of True but with a different case ('On', 'TRUE').