Add config for new bug relationships
Adds a new global config $g_default_bug_relationship, which presets the
relation type in the selection list.
Defaults to BUG_RELATED.

This slightly changes the behavior compared to earlier versions, as the
default was BUG_REL_ANY (-1); however, since this is not a valid entry
in the selection list, the first value (BUG_DEPENDANT) was used instead.