Fallback to English when string is not defined
MantisBT generally falls back correctly to default English when a string
is not defined in the current language. However, when the admin wants to
restrict the available languages ($g_language_choices_arr) and remove
'english' from the list (see issue #13315), an error 300 is displayed
when strings are missing from the translated strings_xxx.txt file (as
can be the case for strings defined as "optional" in TranslateWiki, such
as $s_directionality).

Commit also removes obsolete comment about not falling back to English.