Perform obsolete check against db config also
Previously the check for usage of obsolete configuration variables was
only performed against global variables (defined in config_inc.php).
With this change, a more thorough verification is done, including
values with an override in mantis_config_table.

Detailed information about where the config option is set (database
and/or config_inc.php), for which user(s) and for which project(s), is
printed with the warning message.

To implement this functionality, it was required to alter the order in
which the checks are executed by check/index.php: the DB checks must be
performed before the Configuration checks, otherwise there is no DB
connectivity to verify values in the config table.

Fixes #13435