Allow clearing fixed in/target version for multiple bugs
Prior to this, trying to set the "Fixed in version" or "Target version"
fields to blank for multiple bugs at once (from the "View Issues" page)
failed with error "You did not have appropriate permissions to perform
that action", even when executed as admin. Problem does not occur when
changing to a non-empty value.

Fixes #11801