Allow mass-assign of issues to Updaters
The code handling assignment in bug_actiongroup.php had an incorrect
check for the access level of the user the issues are being assigned
to. It was checked against $t_threshold, which is the minimum access
level required of the current user to perform the action, instead of
the access level to be assigned issues to (handle_bug_threshold).

This affects MantisBT instances where Updaters (or aother role) is
allowed to handle issues (through custom Workflow Thresholds)

Fixes #12324