Correct display custom status in cross-project relationships
In the bug relationships section in view.php, the target bug's status
was incorrectly displayed as "@X@" (where X is the status_enum id) if
the following conditions were true:
 - the target bug is in a different project
 - it is a custom status

This commit fixes the behavior, and displays the corresponding enum
element's language string.

2 new parameters have been added to function get_enum_element(): user
and project id (defaulted to null to preserve existing behavior). These
parameters are passed on to config_get() call, allowing to specify the
context for retrieving the enum element.

The call to get_enum_element() in function relationship_get_details()
has been altered to specify the target bug's project id.

Fixes #11323