Inconsistent config_eval behavior with global configs
When config_eval() is called from config_get_global() to resolve
a recursive config definition, it should call config_get_global() in
return.

This caused incorrect highlighting of some settings as being overriden
in the Workflow Thresholds page, more specifically those configs
defined as having the same value as another one, e.g.
$g_update_bug_assign_threshold = '%handle_bug_threshold%'

The problem was fixed by adding an optional parameter to config_eval()
to specify if it is called in a global context or not.

Fixes #13683