Add Account Unlock button to manage_user_edit_page.php
The button is displayed when it is not possible to reset the user's
password as defined by custom function auth_can_change_password(),
$g_max_failed_login_count > 0 and the user's failed login count is
higher than this threshold.

Prior to this, in the case where it is not possible to reset a user's
password and $g_max_failed_login_count > 0, the administrator's only
way to reset the failed login count was direct SQL update. This was
introduced by commit 3eaa71f7711be41610fedb2d53506483261d8351.

Fixes #13690