Add Account Unlock button to manage_user_edit_page.php
The button is displayed when it is not possible to reset the user's
password as defined by custom function auth_can_change_password(),
$g_max_failed_login_count > 0 and the user's failed login count is
higher than this threshold.

Prior to this, in the case where it is not possible to reset a user's
password and $g_max_failed_login_count > 0, the administrator's only
way to reset the failed login count was direct SQL update. This was
introduced by commit aa48e0c7547cd0edc4df95ff28f67d6f7b3d36a5.

Fixes #13690