Fix SQL errors with SQL Server
This commit addresses 2 occurences of SQL Server error: "Operand type
clash: int is incompatible with text" which are triggered when calling
the following functions:

 - config_set()
 - token_create()
 - token_update()

This is due to the field "Value" in the relevant Mantis table being of
type TEXT, while the corresponding value in the array passed in call to
db_query_bound() is not a string.

Thanks to genius_p for the original patch.

Fixes #12081, #12082