Inconsistent definition of 'relation' $g_notify_flags
In config_defaults_inc.php, the action for $g_notify_flags to alert
when relationships change is defined as 'relation'. This is the value
used by the email api. However, the manage_config_email_page.php used
a different value 'relationship' to store/retrieve the configuration.

This leads to inconsistent behavior: when defined in config_inc.php
with 'relation', the flags are set as expected, but not reflected in
the UI. When changed in the UI, the settings have no effect (e.g.
e-mails are sent even though the flag says they are disabled).

Fixes #13714