Prevent unwanted clearing version fields when editing bugs
When editing issues having Version information set to some value that is
no longer available in the selection list (i.e. because the version has
been marked as obsolete), Mantis defaulted the corresponding Version
field to blank in the bug_update_page. If the user then saved the bug,
the version information was consequently cleared in the DB, possibly
without the user even noticing the change.

This commit ensures that the version selection list always includes the
currently set version to avoid this behavior.

Fixes #13096