Adds support for date format d/m/Y
When the date format is set to 'd/m/Y', the php strtotime function does not
work correctly, it will try to parse the string as m/d/Y instead. The new
function 'strtotime_safe' will check the $g_short_date_format setting,
and if set to 'd/m/Y' it will fix this.