Avoid error when editing bug with invalid version
When editing an issue having the Product Version field set to a version
number which is not defined in the issue's current project (which can
happen when the issue has been moved from another project, or a version
has been deleted from the current project), APPLICATION ERROR 1601 -
Version "" not found occurs.

The error occurs because the version field in the bug table is not
cleared when the issue is moved, so when Mantis is building the version
selection list to edit the bug, it tries to add the version defined in
it to the list, which triggers the error since that version does not
exist in the current project.

Regression caused by commit a68398245663d25229800638b6965ffe6e7ab0d3

Fixes #13938