Extra column in csv export when sponsorship_total >= 1000
Bug introduced by commit 169b882b97b0c6cf1babbf072ce326f0779c6c0b.

Due to the use of number_format() function with default parameters, PHP
outputs the sponsorship amount formatted with ',' as a thousands
separator, which also happens to be MantisBT's default csv separator.
Since the fields are not escaped, this causes an extra column to be
generated.

Call to number_format() was removed, as it makes more sense to output
raw data in csv exports.

Fixes #13958